int -> bigint

+++ tools/serverpod_cli/lib/src/generator/types.dart
@@ -174,7 +174,7 @@ class TypeDefinition {
     // TODO: add all suported types here
     if (className == 'String') return 'text';
     if (className == 'bool') return 'boolean';
-    if (className == 'int' || isEnum) return 'integer';
+    if (className == 'int' || isEnum) return 'bigint';
     if (className == 'double') return 'double precision';
     if (className == 'DateTime') return 'timestamp without time zone';
     if (className == 'ByteData') return 'bytea';
